third_party/WebKit/LayoutTests/fast/viewport/viewport-78-expected.txt

CONSOLE ERROR: line 3: The key "a" is not recognized and ignored.
CONSOLE ERROR: line 3: The key "1" is not recognized and ignored.
ALERT: viewport size 100x110 scale 3.2 with limits [3.2, 5] and userScalable true
